Refactor image set selection & processing
The code had a confusing structure and it was difficult to tell the outputs of
the imageSetExtraction.
The new structure has imageSetExtraction as a script and imageSetSelection and
imageSetProcessing as functions. The output of the two functions are the only
variables that are put into the workspace.

Further work should be done to differentiate between image set methods and
functions and between similar names such as imageSetExtraction &
imageSetProcessing.