Fix calibGUIWindow variable not in workspace
The variable I created to keep track of the gui window was not going into the
workspace and therefore could not be called by my method. I solved this by
forcing the variable to be saved in the workspace with the assignin function.

This problem seems to have occurred because MATLAB handles a script directly
called by a function as part of the function's workspace, therefore the
variables created by the script were hidden similarly to a variable in a
function. I did not test this theory whatsoever.