Refactor stereoCalibPreprocess
The preprocess process was difficult to read and had repeated code, so I made
it so that the script starts with setting the active variables then going
though an individual calibPreprocess. This makes the stereoCalibPreprocess
easier to read, but the calibPreprocess uses several original toolbox methods
so it's difficult to read.